unsigned int decimal

+++ libftprintf.h
@@ -4,11 +4,11 @@
 
 int		ft_putstr(char *strtoprint, int prec);
 int		ft_printf(const char *s, ...);
-int		ft_intlen(int nbr);
+int		ft_intlen(long nbr);
 int		ft_putchar(int n);
 
 int		ft_puthex(long n);
-int		ft_putdec(int n, int prec);
+int		ft_putdec(long n, int prec);
 void	foo(char *fmt, ...);
 
 #endif	/* LIBFTPRINTF_H */
